add tsconfig and console API types
Figma plugins run in a restricted environment where the browser APIs
aren't available by default, including 'console'.  This means the plugin
typings should declare exactly what console API is available.  (Note that
the functions listed here are a subset of the full browser console API;
these are the subset that plugins actually have access to.)

This change also adds a tsconfig.json to this repo, because otherwise
TypeScript thinks this definition of console conflicts with the ambient
browser one.